requirements.txt
fastapi[all]
requests
python-dotenv
jinja2